[deux] <Todo> Newlib porting

+++ deux/ports_libs/libc/argz/argz_count.c
@@ -4,6 +4,7 @@
  * is freely granted, provided that this notice is preserved.
  */
 
+#include <_ansi.h>
 #include <argz.h>
 #include <stddef.h>
 #include <sys/types.h>
@@ -12,7 +13,7 @@ size_t
 argz_count (const char *argz,
        size_t argz_len)
 {
-  size_t i;
+  int i;
   size_t count = 0;
 
   for (i = 0; i < argz_len; i++)
